Stream the unikernel image to albatross
Instead of fully reading and writing it to albatross, allow there to be a
chunked stream. So, the unikernel image is written as data on the TLS flow in
a chunked encoding (4 byte length followed by the chunk -- end of the data is
marked by a 0 chunk).

Relies on https://github.com/robur-coop/albatross/pull/211, still needs the
unikernel image to be streamed from the HTTP request to albatross / across the
request handler.